+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/blocks/AbstractCandleSkullBlock.java
@@ -28,6 +28,7 @@ import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
@@ -206,7 +207,7 @@ public abstract class AbstractCandleSkullBlock extends AbstractCandleBlock imple
 
     @Override
     public boolean canBeExtinguishedBy(ItemStack item) {
-        return item.isEmpty() || ILightable.super.canBeExtinguishedBy(item);
+        return  item.isEmpty() || ILightable.super.canBeExtinguishedBy(item);
     }
 
     @Override